Recursive traversal of Service method args and return type for marshal/unmarshal code.
This commit includes more robust handling of Messages with fields that are Messages when
determining which Messages should get JSONTo and ToJSON functions.  The initial solution
didn't account for this, so functions weren't being generated even though they were actually
required.

This isn't the best solution yet since it still requires specific knowledge of when to call methods on APIContext.